ファイル同士をマージ　機能することを確認済み
